+++ src/http/representation/ResourceIdentifier.ts
@@ -9,8 +9,15 @@ export interface ResourceIdentifier {
 }
 
 /**
- * Determines whether the object is a `ResourceIdentifier`.
+ * Determines whether the object is a {@link ResourceIdentifier}.
  */
 export function isResourceIdentifier(object: any): object is ResourceIdentifier {
   return object && (typeof object.path === 'string');
+}
+
+/**
+ * Factory function creating a {@link ResourceIdentifier} for convenience.
+ */
+export function createResourceIdentifier(resourcePath: string): ResourceIdentifier {
+  return { path: resourcePath };
 }