+++ src/storage/DataAccessorBasedStore.ts
@@ -6,7 +6,7 @@ import type { AuxiliaryStrategy } from '../http/auxiliary/AuxiliaryStrategy';
 import { BasicRepresentation } from '../http/representation/BasicRepresentation';
 import type { Patch } from '../http/representation/Patch';
 import type { Representation } from '../http/representation/Representation';
-import type { RepresentationMetadata } from '../http/representation/RepresentationMetadata';
+import { RepresentationMetadata } from '../http/representation/RepresentationMetadata';
 import type { ResourceIdentifier } from '../http/representation/ResourceIdentifier';
 import { getLoggerFor } from '../logging/LogUtil';
 import { INTERNAL_QUADS } from '../util/ContentTypes';
@@ -39,10 +39,12 @@ import {
   SOLID_META,
   PREFERRED_PREFIX_TERM,
   CONTENT_TYPE_TERM,
+  SOLID_AS,
+  AS,
 } from '../util/Vocabularies';
 import type { DataAccessor } from './accessors/DataAccessor';
 import type { Conditions } from './Conditions';
-import type { ResourceStore } from './ResourceStore';
+import type { ResourceStore, ChangeMap } from './ResourceStore';
 
 /**
  * ResourceStore which uses a DataAccessor for backend access.
@@ -138,7 +140,7 @@ export class DataAccessorBasedStore implements ResourceStore {
   }
 
   public async addResource(container: ResourceIdentifier, representation: Representation, conditions?: Conditions):
-  Promise<ResourceIdentifier> {
+  Promise<ChangeMap> {
     this.validateIdentifier(container);
 
     const parentMetadata = await this.getSafeNormalizedMetadata(container);
@@ -174,13 +176,11 @@ export class DataAccessorBasedStore implements ResourceStore {
     }
 
     // Write the data. New containers should never be made for a POST request.
-    await this.writeData(newID, representation, isContainer, false, false);
-
-    return newID;
+    return this.writeData(newID, representation, isContainer, false, false);
   }
 
   public async setRepresentation(identifier: ResourceIdentifier, representation: Representation,
-    conditions?: Conditions): Promise<ResourceIdentifier[]> {
+    conditions?: Conditions): Promise<ChangeMap> {
     this.validateIdentifier(identifier);
 
     // Check if the resource already exists
@@ -216,7 +216,7 @@ export class DataAccessorBasedStore implements ResourceStore {
   }
 
   public async modifyResource(identifier: ResourceIdentifier, patch: Patch,
-    conditions?: Conditions): Promise<ResourceIdentifier[]> {
+    conditions?: Conditions): Promise<never> {
     if (conditions) {
       let metadata: RepresentationMetadata | undefined;
       try {
@@ -233,7 +233,7 @@ export class DataAccessorBasedStore implements ResourceStore {
     throw new NotImplementedHttpError('Patches are not supported by the default store.');
   }
 
-  public async deleteResource(identifier: ResourceIdentifier, conditions?: Conditions): Promise<ResourceIdentifier[]> {
+  public async deleteResource(identifier: ResourceIdentifier, conditions?: Conditions): Promise<ChangeMap> {
     this.validateIdentifier(identifier);
     const metadata = await this.accessor.getMetadata(identifier);
     // Solid, §5.4: "When a DELETE request targets storage’s root container or its associated ACL resource,
@@ -266,22 +266,25 @@ export class DataAccessorBasedStore implements ResourceStore {
     // Solid, §5.4: "When a contained resource is deleted,
     // the server MUST also delete the associated auxiliary resources"
     // https://solid.github.io/specification/protocol#deleting-resources
-    const deleted = [ identifier ];
+    const changes: ChangeMap = {};
     if (!this.auxiliaryStrategy.isAuxiliaryIdentifier(identifier)) {
       const auxiliaries = this.auxiliaryStrategy.getAuxiliaryIdentifiers(identifier);
-      deleted.push(...await this.safelyDeleteAuxiliaryResources(auxiliaries));
+      for (const deletedId of await this.safelyDeleteAuxiliaryResources(auxiliaries)) {
+        changes[deletedId.path] = this.createActivityMetadata(deletedId, AS.Delete);
+      }
     }
 
     if (!this.identifierStrategy.isRootContainer(identifier)) {
       const container = this.identifierStrategy.getParentContainer(identifier);
-      deleted.push(container);
+      changes[container.path] = this.createActivityMetadata(container, AS.Update);
 
       // Update modified date of parent
       await this.updateContainerModifiedDate(container);
     }
 
     await this.accessor.deleteResource(identifier);
-    return deleted;
+    changes[identifier.path] = this.createActivityMetadata(identifier, AS.Delete);
+    return changes;
   }
 
   /**
@@ -359,7 +362,7 @@ export class DataAccessorBasedStore implements ResourceStore {
    * @returns Identifiers of resources that were possibly modified.
    */
   protected async writeData(identifier: ResourceIdentifier, representation: Representation, isContainer: boolean,
-    createContainers: boolean, exists: boolean): Promise<ResourceIdentifier[]> {
+    createContainers: boolean, exists: boolean): Promise<ChangeMap> {
     // Make sure the metadata has the correct identifier and correct type quads
     // Need to do this before handling container data to have the correct identifier
     representation.metadata.identifier = DataFactory.namedNode(identifier.path);
@@ -382,18 +385,22 @@ export class DataAccessorBasedStore implements ResourceStore {
     // Solid, §5.3: "Servers MUST create intermediate containers and include corresponding containment triples
     // in container representations derived from the URI path component of PUT and PATCH requests."
     // https://solid.github.io/specification/protocol#writing-resources
-    const modified = [];
+    let changes: ChangeMap = {};
     if (!this.identifierStrategy.isRootContainer(identifier) && !exists) {
-      const container = this.identifierStrategy.getParentContainer(identifier);
+      const parent = this.identifierStrategy.getParentContainer(identifier);
       if (!createContainers) {
-        modified.push(container);
+        changes[parent.path] = this.createActivityMetadata(parent, AS.Update);
       } else {
-        const created = await this.createRecursiveContainers(container);
-        modified.push(...created.length === 0 ? [ container ] : created);
+        const createdContainers = await this.createRecursiveContainers(parent);
+        changes = { ...changes, ...createdContainers };
+
+        if (Object.keys(createdContainers).length === 0) {
+          changes[parent.path] = this.createActivityMetadata(parent, AS.Update);
+        }
       }
 
       // Parent container is also modified
-      await this.updateContainerModifiedDate(container);
+      await this.updateContainerModifiedDate(parent);
     }
 
     // Remove all generated metadata to prevent it from being stored permanently
@@ -403,7 +410,8 @@ export class DataAccessorBasedStore implements ResourceStore {
       this.accessor.writeContainer(identifier, representation.metadata) :
       this.accessor.writeDocument(identifier, representation.data, representation.metadata));
 
-    return [ ...modified, identifier ];
+    changes[identifier.path] = this.createActivityMetadata(identifier, exists ? AS.Update : AS.Create);
+    return changes;
   }
 
   /**
@@ -597,7 +605,7 @@ export class DataAccessorBasedStore implements ResourceStore {
    * Will throw errors if the identifier of the last existing "container" corresponds to an existing document.
    * @param container - Identifier of the container which will need to exist.
    */
-  protected async createRecursiveContainers(container: ResourceIdentifier): Promise<ResourceIdentifier[]> {
+  protected async createRecursiveContainers(container: ResourceIdentifier): Promise<ChangeMap> {
     // Verify whether the container already exists
     try {
       const metadata = await this.getNormalizedMetadata(container);
@@ -609,7 +617,7 @@ export class DataAccessorBasedStore implements ResourceStore {
       if (!isContainerPath(metadata.identifier.value)) {
         throw new ForbiddenHttpError(`Creating container ${container.path} conflicts with an existing resource.`);
       }
-      return [];
+      return {};
     } catch (error: unknown) {
       if (!NotFoundHttpError.isInstance(error)) {
         throw error;
@@ -618,9 +626,14 @@ export class DataAccessorBasedStore implements ResourceStore {
 
     // Create the container, starting with its parent
     const ancestors = this.identifierStrategy.isRootContainer(container) ?
-      [] :
+      {} :
       await this.createRecursiveContainers(this.identifierStrategy.getParentContainer(container));
-    await this.writeData(container, new BasicRepresentation([], container), true, false, false);
-    return [ ...ancestors, container ];
+    const changes = await this.writeData(container, new BasicRepresentation([], container), true, false, false);
+
+    return { ...changes, ...ancestors };
+  }
+
+  private createActivityMetadata(id: ResourceIdentifier, activity: string): RepresentationMetadata {
+    return new RepresentationMetadata(id, { [SOLID_AS.terms.Activity.value]: activity });
   }
 }